pyproject.toml
[tool.poetry]
name = "code-tldr"
version = "0.1.0"
description = "Code TLDR"
authors = ["khw1031 <khw1031@gmail.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = ">=3.13"
fastapi = ">=0.115.12,<0.116.0"
boto3 = ">=1.37.27,<2.0.0"
python-dotenv = ">=1.1.0,<2.0.0"
pydantic = ">=2.11.2,<3.0.0"
pydantic-settings = ">=2.8.1,<3.0.0"
uvicorn = ">=0.34.0,<0.35.0"
atlassian-python-api = "^4.0.3"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dev = "app.main:start_dev"
start = "app.main:start_prod"

[tool.black]
line-length = 88
target-version = ['py313']
skip-string-normalization = true